updated uints

+++ contracts/Cointinuum.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.17;
 
 //--------------------------------------
-//  Cointinuum: CTM ERC20 Token contract
+//  Cointinuum: ERC20 Token contract
 //
 // Symbol      : CTM
 // Name        : Cointinuum
@@ -14,14 +14,14 @@ contract Cointinuum {
 
     bytes32 public name;
     bytes32 public symbol;
-    uint8 public decimals;
+    uint256 public decimals; // Looking into changing this value to a lower uint value for saving gas
     uint256 private initialSupply;
     uint256 public _totalSupply;
     address private owner;
-    uint256 public constant MAX_TOTAL_SUPPLY = 110000000; // MAX SUPPLY SET CONST
+    uint256 public constant MAX_TOTAL_SUPPLY = 110000000; // MAX SUPPLY CONST
 
-    mapping(address => uint) balances;
-    mapping(address => mapping(address => uint)) allowed;
+    mapping(address => uint256) balances;
+    mapping(address => mapping(address => uint256)) allowed;
 
     event Transfer(address indexed from, address indexed to, uint256 value);
     event Approval(address indexed owner, address indexed spender, uint256 value);
@@ -38,15 +38,15 @@ contract Cointinuum {
         owner = msg.sender;
     }
 
-    function totalSupply() public view returns (uint) {
+    function totalSupply() public view returns (uint256) {
         return _totalSupply;
     }
 
-    function balanceOf(address tokenOwner) public view returns (uint balance) {
+    function balanceOf(address tokenOwner) public view returns (uint256 balance) {
         return balances[tokenOwner];
     }
 
-    function approve(address spender, uint tokens) public returns (bool) {
+    function approve(address spender, uint256 tokens) public returns (bool) {
         require(spender != address(0));
         require(tokens <= _totalSupply);
 
@@ -55,7 +55,7 @@ contract Cointinuum {
         return true;
     }
 
-    function allowance(address tokenOwner, address spender) public view returns (uint remaining) {
+    function allowance(address tokenOwner, address spender) public view returns (uint256 remaining) {
         return allowed[tokenOwner][spender];
     }
 
@@ -70,7 +70,7 @@ contract Cointinuum {
         return true;
     }
 
-    function transferFrom(address from, address to, uint tokens) public returns (bool) {
+    function transferFrom(address from, address to, uint256 tokens) public returns (bool) {
         require(to != address(0));
         require(tokens <= balances[from]);
         require(tokens <= allowed[from][msg.sender]);
